add savepoint/rollback/commit to MemoryTableFactory

+++ libstorage/LevelDBStorage.h
@@ -41,7 +41,7 @@ public:
     virtual Entries::Ptr select(
         h256 hash, int num, const std::string& table, const std::string& key) override;
     virtual size_t commit(
-        h256 hash, int num, const std::vector<TableData::Ptr>& datas, h256 blockHash) override;
+        h256 hash, int64_t num, const std::vector<TableData::Ptr>& datas, h256 blockHash) override;
     virtual bool onlyDirty() override;
 
     void setDB(std::shared_ptr<leveldb::DB> db);
