+++ libstorage/Storage.h
@@ -46,7 +46,7 @@ public:
     virtual Entries::Ptr select(
         h256 hash, int num, const std::string& table, const std::string& key) = 0;
     virtual size_t commit(
-        h256 hash, int num, const std::vector<TableData::Ptr>& datas, h256 blockHash) = 0;
+        h256 hash, int64_t num, const std::vector<TableData::Ptr>& datas, h256 blockHash) = 0;
     virtual bool onlyDirty() = 0;
 };
